Extract WavReader to own file and move it to FileFormats
